fix: grid layout, card styling, pageSize pagination filters, page/view separation
