+++ app/src/components/common/icons/IconArrowBack.tsx
@@ -19,10 +19,11 @@ interface Props {
   hoverEffect?: boolean
   onClick?: () => void
   style?: any
+  color?: string
 }
 
 export const IconArrowBack = (props: Props) => {
-  const { hoverEffect = false, ...restProps } = props
+  const { color, hoverEffect = false, ...restProps } = props
 
   return (
     <Wrapper
@@ -37,7 +38,7 @@ export const IconArrowBack = (props: Props) => {
       <path
         className="path"
         d="M21 11H6.83L10.41 7.41L9 6L3 12L9 18L10.41 16.59L6.83 13H21V11Z"
-        fill={hoverEffect ? '#DCDFF2' : '#37474F'}
+        fill={color ? color : hoverEffect ? '#DCDFF2' : '#37474F'}
       />
     </Wrapper>
   )
