--- app/src/components/common/icons/IconBack.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const IconBack = () => (
-  <svg fill="none" height="12" viewBox="0 0 18 12" width="18" xmlns="http://www.w3.org/2000/svg">
-    <path d="M18 5H3.83L7.41 1.41L6 0L0 6L6 12L7.41 10.59L3.83 7H18V5Z" />
-  </svg>
-)
