+++ app/src/components/common/icons/index.ts
@@ -42,4 +42,3 @@ export { IconFail } from './IconFail'
 export { IconInfo } from './IconInfo'
 export { IconPokt } from './IconPokt'
 export { IconXdai } from './IconXdai'
-export { IconBack } from './IconBack'
