+++ app/src/components/market/market_list/market_home.tsx
@@ -260,6 +260,7 @@ export const MarketHome: React.FC<Props> = (props: Props) => {
   const [currency, setCurrency] = useState<Maybe<string> | null>(currentFilter.currency)
   const [templateId, setTemplateId] = useState<Maybe<string>>(currentFilter.templateId)
   const [curationSource, setCurationSource] = useState<CurationSource>(currentFilter.curationSource)
+  const { first } = currentFilter
 
   const advancedFilterSelectedCount = [currency, arbitrator, curationSource !== CurationSource.ALL_SOURCES].filter(
     element => element,
@@ -351,6 +352,7 @@ export const MarketHome: React.FC<Props> = (props: Props) => {
 
   useEffect(() => {
     onFilterChange({
+      first,
       arbitrator,
       curationSource,
       templateId,
@@ -362,6 +364,7 @@ export const MarketHome: React.FC<Props> = (props: Props) => {
       title,
     })
   }, [
+    first,
     arbitrator,
     curationSource,
     templateId,
